nvtable: refactor - moving NVHandleDesc to a separate header
